Refactor metadata extractor
The previous implementation took the container and streams of an input, combined their metadata into a single map, and indexed into values for keys like "title", "artist", and "album". This wouldn't work for formats with alternative naming (the only normalization performed was lowercasing all keys, which was a bad idea, given metadata keys are case-sensitive). The new implementation relies on the container and its first stream, combines the metadata into a single map, and maps said map to a new map normalizing the keys. For example, "title" and "TITLE" map to "title". The result is any name can be mapped to a common name (say, "TOTALTRACKS" to "TRACKTOTAL", although we don't perform such an action).