Display decoded cover art for album
The previous implementation read the attached picture packet and relied on the system's built-in decoders to display the compressed image. This implementation uses FFmpeg to decode and scale (i.e. transform the color space) the image, feeding the result into CGImage.

This process likely makes no practical difference to end users, as cover arts are overwhelmingly in PNG or JPEG; however, the code is now in place to write the audio file decoder.