Support AIFF, Vorbis, Wave64 & WavPack
This just about completes the codecs of interest to be bundled into the app. Direct Stream Digital (DSD) is an option, but from the spectrogram ffprobe generated from an XLD encode, it looks like either XLD can't encode it or FFmpeg can't decode it appropriately.

In addition, builds have been refactored to support universal compilation. This is a step up from performing builds for each architecture (x86 and ARM). In the future, when iOS is developed, the cross platform options may be necessary.