Reorganize build scripts
This builds Ogg, Vorbis, and Opus as fat (i.e., universal) binaries, while keeping FFmpeg as an architecture-dependent build (x86_64 or arm64). The benefit is that release builds should eventually be faster from the shorter lifecycle.

In addition, the file structure has been reorganized to make Xcode an external input to the system. That is, Xcode run scripts set variables that'll be transformed to fit the domain of the build scripts, meaning less complexity.

Finally, we added back `make distclean` because it's necessary for some configures to pick up changes.