Reorganize scripts
Xcode scripts are prefixed to signal to the user that they shouldn't run them themselves.

Xcode run scripts use the top-level Tools directory as its input path so we don't have to maintain a set of paths. This does mean that the run script will be executed more often, but only when changes are made to the scripts, which is not the worst cost in the world.